+++ packages/ui/src/atoms/Select.tsx
@@ -69,11 +69,7 @@ const SelectTrigger = React.forwardRef<
   >
     {children}
     <SelectPrimitive.Icon asChild={true}>
-      <Icon
-        name="chevron-down"
-        size={16}
-        className="text-grey-600 dark:text-grey-400"
-      />
+      <Icon name="chevron-down" className="size-4 text-grey-600 dark:text-grey-400" />
     </SelectPrimitive.Icon>
   </SelectPrimitive.Trigger>
 ));
@@ -90,7 +86,7 @@ const SelectScrollUpButton = React.forwardRef<
     className={cn("flex cursor-default items-center justify-center py-1", className)}
     {...props}
   >
-    <Icon name="chevron-down" size={16} className="text-grey-600" />
+    <Icon name="chevron-down" className="size-4 text-grey-600" />
   </SelectPrimitive.ScrollUpButton>
 ));
 SelectScrollUpButton.displayName = SelectPrimitive.ScrollUpButton.displayName;
@@ -106,7 +102,7 @@ const SelectScrollDownButton = React.forwardRef<
     className={cn("flex cursor-default items-center justify-center py-1", className)}
     {...props}
   >
-    <Icon name="chevron-down" size={16} className="text-grey-600 dark:text-grey-400" />
+    <Icon name="chevron-down" className="size-4 text-grey-600 dark:text-grey-400" />
   </SelectPrimitive.ScrollDownButton>
 ));
 SelectScrollDownButton.displayName = SelectPrimitive.ScrollDownButton.displayName;
@@ -174,7 +170,7 @@ const SelectItem = React.forwardRef<
   >
     <span className="absolute right-2 flex h-3.5 w-3.5 items-center justify-center">
       <SelectPrimitive.ItemIndicator>
-        <Icon name="check" size={16} className="text-grey-600 dark:text-grey-400" />
+        <Icon name="check" className="size-4 text-grey-600 dark:text-grey-400" />
       </SelectPrimitive.ItemIndicator>
     </span>
 
@@ -275,6 +271,7 @@ const Select = React.forwardRef<
           className={className}
           size={size}
           autoFocus={autoFocus}
+          translate="no"
         >
           <SelectValue placeholder={placeholder} />
         </SelectTrigger>