+++ packages/ui/src/index.ts
@@ -6,8 +6,9 @@ export * from "./molecules";
 /* utils */
 export * from "./styles/screens";
 export * from "./utils/DesignSystemContext";
-export * from "./utils/useBreakpoints";
 export * from "./utils/cn";
+export * from "./utils/react";
+export * from "./utils/useBreakpoints";
 
 export const VERSION = import.meta.env.VITE_VERSION;
 export const BUILD_INFO = import.meta.env.VITE_BUILD_INFO;